+++ src/main/java/Dominio/ContaEspecial.java
@@ -3,10 +3,12 @@ package Dominio;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class ContaEspecial extends Conta {
 
-  private Float limite;
+  private BigDecimal limite;
 
 }
